estrututra do produto finalizado

+++ transport_logistics/src/main/java/br/com/senior/transport_logistics/domain/product/dto/response/ProductResponseDTO.java
@@ -8,10 +8,10 @@ public record ProductResponseDTO(
         long id,
         String name,
         ProductCategory category,
-        float weight
+        double weight
 ) {
 
-    public ProductResponseDTO(long id, String name, ProductCategory category, float weight) {
+    public ProductResponseDTO(long id, String name, ProductCategory category, double weight) {
         this.id = id;
         this.name = name;
         this.category = category;
